Delete unused email components and related API routes
