+++ src/components/email/BaseLayout.tsx
